perf(apu_micrones): inline timer_advance, optimize DC tracker, shrink hot counters
Three targeted APU hot-path improvements:

1. apu_timer_advance: static inline instead of MICRONES_HOT_FUNC
   MICRONES_HOT_FUNC forces the function into a named .iram1.text section,
   which prevents the compiler from inlining it.  Since the callers
   (apu_step's inner loop) are already in IRAM, inlining is strictly better:
   the code ends up in IRAM with zero CALL8/ENTRY/RETW overhead.
   ~3200 calls/frame × ~5 cycles each → ~73 µs saved at 240 MHz.

2. apu_mix_sample: eliminate float round-trip in DC tracker
   Old:  mixed = mixed_raw / 32767;
         dc   += (mixed - dc) * 0.0001;
         sample = lrintf((mixed - dc) * 32767)
   New:  dc_level_tracker stores dc×32767 (scaled domain)
         dc   += (mixed_raw - dc) * 0.0001;
         sample = lrintf(mixed_raw - dc)
   Eliminates 2 float multiplies per sample × 800 samples/frame → ~27 µs.

3. cpu_cycles, frame_counter_cycle, sample_count: uint64_t → uint32_t
   All three are accessed every inner-loop iteration of apu_step.
   cpu_cycles wraps after ~40 min (harmless; only parity used).
   frame_counter_cycle resets every ~14915 APU cycles; uint32_t is ample.
   sample_count overflows after ~24 h; fine for embedded use.
   Eliminates 64-bit add/load/store pairs → ~17 µs combined.

Combined expected saving: ~117 µs/frame, bringing the target from ~17000 µs
closer to the 16667 µs (60 fps) budget.

+++ src/common/apu_micrones.h
@@ -166,11 +166,11 @@ typedef struct {
     uint64_t register_write_count[APU_DEBUG_REGISTER_COUNT];
     uint8_t mix_enable_mask;
     uint8_t test_tone_mode;
-    uint64_t cpu_cycles;
-    uint64_t sample_count;
+    uint32_t cpu_cycles;         /* low 32 bits sufficient; wraps harmlessly after ~40 min */
+    uint32_t sample_count;       /* uint32_t overflows after ~24 h; fine for embedded use */
     uint64_t dropped_samples;
     uint64_t clip_count;
-    uint64_t frame_counter_cycle;
+    uint32_t frame_counter_cycle; /* resets every ~14915 APU cycles; uint32_t is ample */
     uint64_t frame_counter_steps;
     uint32_t sample_phase;
     uint32_t pcm_read_index;
